move crawler funcions to collector
