+++ .gitignore
@@ -1,1 +1,7 @@
 .DS_Store
+**__pycache__**
+collectors/secrets/*
+collectors/webdrivers/*
+collectors/data/csv/*
+collectors/data/raw/*
+**/**.csv
