--- cloud-functions/functions/gcp_cloud_scraper/gcp_cloud_scraper/items.py
@@ -1,12 +1,0 @@
-# Define here the models for your scraped items
-#
-# See documentation in:
-# https://docs.scrapy.org/en/latest/topics/items.html
-
-import scrapy
-
-
-class GcpCloudScraperItem(scrapy.Item):
-    # define the fields for your item here like:
-    # name = scrapy.Field()
-    pass
