+++ collectors/README.md
@@ -1,0 +1,17 @@
+### Data Collectors 
+
+* Not running on cloud functions
+* run on local macOS
+  * developed on mac air M2 ARM
+
+
+* To get data ( for gemini RAG langchains )
+```
+cd parsers
+
+python3 aws.py
+python3 gcp.py
+python3 azure.py # not working yet
+
+
+```
