+++ collectors/parsers/__init__.py
@@ -1,0 +1,3 @@
+import sys
+import os
+sys.path.append(os.path.join(os.path.dirname(__file__), "../../"))
