+++ collectors/pyproject.toml
@@ -1,0 +1,33 @@
+[tool.poetry]
+name = "data-collectors"
+version = "0.1.0"
+description = ""
+authors = ["kade <ydh0924@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+google-cloud-storage = "^2.18.0"
+functions-framework = "==3.*"
+beautifulsoup4 = "^4.12.3"
+google-auth = "^2.32.0"
+google-auth-oauthlib = "^1.2.1"
+google-auth-httplib2 = "^0.2.0"
+google-api-python-client = "^2.138.0"
+langchain = "^0.2.11"
+langchain-community = "^0.2.10"
+langchain-google-genai = "^1.0.8"
+python-dotenv = "^1.0.1"
+gitpython = "^3.1.43"
+pydantic = "^2.8.2"
+google-cloud-firestore = "^2.17.0"
+selenium = "^4.23.1"
+webdriver-manager = "^4.0.2"
+pytz = "^2024.1"
+geckodriver-autoinstaller = "^0.1.0"
+pandas = "^2.2.2"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
